requirements.txt
fastapi
uvicorn
python-dotenv
gspread
oauth2client
jinja2
pydantic
streamlit
streamlit-authenticator
python-multipart